refactor(checkbox): merge MultiCheckbox logic into InputCheckbox with support for options[]
