+++ frontend/src/ui/inputs/checkbox/InputCheckbox.jsx
@@ -1,49 +1,53 @@
-import { useState } from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
 
-const InputCheckbox = ({
-  id,
-  name,
-  label,
-  onChange = () => {},
-  defaultChecked = false,
-  disabled = false,
-  required = false
-}) => {
-  const [isChecked, setIsChecked] = useState(defaultChecked);
 
-  if (disabled && !defaultChecked) required = false;
+const InputCheckbox = ({ options = [], onChange = () => {} }) => {
+  const [selectedIds, setSelectedIds] = useState(
+    options.filter(o => o.defaultChecked).map(o => o.id)
+  );
 
-  const handleCheckboxChange = (event) => {
-    const checked = event.target.checked;
-    setIsChecked(checked);      
-    onChange(event);              
+  const handleCheckboxChange = (id, isChecked) => {
+    const updated = isChecked
+      ? [...selectedIds, id]
+      : selectedIds.filter(item => item !== id);
+
+    setSelectedIds(updated);
+    onChange(updated); 
   };
 
   return (
-    <label htmlFor={id} className="flex gap-2 items-center">
-      <input
-        type="checkbox"
-        id={id}
-        name={name}
-        checked={isChecked}
-        onChange={handleCheckboxChange}
-        disabled={disabled}
-        required={required}
-      />
-      <span>{label}</span>
-    </label>
+    <div className="flex flex-col gap-2">
+      {options.map(({ id, name, label, disabled = false, required = false, defaultChecked = false }) => (
+        <label key={id} htmlFor={id} className="flex gap-2 items-center">
+          <input
+            type="checkbox"
+            id={id}
+            name={name}
+            defaultChecked={defaultChecked}
+            disabled={disabled}
+            required={required}
+            onChange={e => handleCheckboxChange(id, e.target.checked)}
+          />
+          <span>{label}</span>
+        </label>
+      ))}
+    </div>
   );
 };
 
 InputCheckbox.propTypes = {
-  id: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  label: PropTypes.string.isRequired,
-  onChange: PropTypes.func,
-  defaultChecked: PropTypes.bool,
-  disabled: PropTypes.bool,
-  required: PropTypes.bool
+  options: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+      label: PropTypes.string.isRequired,
+      disabled: PropTypes.bool,
+      required: PropTypes.bool,
+      defaultChecked: PropTypes.bool
+    })
+  ).isRequired,
+  onChange: PropTypes.func
 };
 
 export default InputCheckbox;
